BioBot_app/requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=15.0
scikit-learn>=1.2
faiss-cpu==1.7.4; platform_system != "Windows"
sentence-transformers>=2.7
fastapi>=0.111
uvicorn[standard]>=0.22
python-dotenv>=1.0
twilio>=9.0
unidecode>=1.3
streamlit>=1.36
